Feat: created 3 examples of atomic pointers using a literal as a server connection and using it passing it as a variable. Used as a atomic pointr i't will show a connection each time it passed the propper pointer. Then another example wich will put to sleep a showConnection implementation every 10 secconds and reach its memory value. This implemented with a channel which will show a boolean and an state. And lastly a data race when two go routines access the same variable and concurrently and at least one of the accesses is a write. To quickly verify a data race, executing go run with race flag to perform a test.
Go run main go main.go race

Go atomic types are an easy way to manage shared resources. it removes the need to constantly implement a mutex to control resources access. This doesn't imply that mutexes are obsolete, as they are still required in opertaions. In conclusion, atomic.Pointer is a great way to bring atomic memory primitives into your program. It is a simple way to prevent data races without fancy mutex code in place. Even if it's only in a little part of a code.